Move AFC CLI code to cli/afc.py
